What does gcc actually want here?

+++ src/wrapper/clap-juce-wrapper.cpp
@@ -1333,7 +1333,7 @@ class ClapJuceWrapper : public clap::helpers::Plugin<
         auto sofar = 0U;
         while (sofar < sz)
         {
-            auto written = stream->write(stream, dat, (uint32_t)((int)sz - sofar));
+            auto written = stream->write(stream, dat, (uint32_t)(sz - sofar));
             if (written < 0)
             {
                 return false;
